requirements.txt
ultralytics
opencv-python
gpiozero
python-telegram-bot
python-dotenv
flask
flask-sqlalchemy
boto3
zstd
psycopg2-binary